# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/toolsets": {
      "post": {
        "summary": "Create Toolset Endpoint",
        "description": "創建新的工具集。\nPOST /api/v1/toolsets",
        "operationId": "create_toolset_endpoint_api_v1_toolsets_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ToolSetCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ToolSet"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "List Toolsets Endpoint",
        "description": "列出所有工具集。\nGET /api/v1/toolsets\nGET /api/v1/toolsets?is_global=true",
        "operationId": "list_toolsets_endpoint_api_v1_toolsets_get",
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "is_global",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Is Global"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ToolSet"
                  },
                  "title": "Response List Toolsets Endpoint Api V1 Toolsets Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/toolsets/{toolset_id}": {
      "get": {
        "summary": "Get Toolset Endpoint",
        "description": "獲取工具集詳情（包含工具列表）。\nGET /api/v1/toolsets/{toolset_id}",
        "operationId": "get_toolset_endpoint_api_v1_toolsets__toolset_id__get",
        "parameters": [
          {
            "name": "toolset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Toolset Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update Toolset Endpoint",
        "description": "更新工具集。\nPUT /api/v1/toolsets/{toolset_id}",
        "operationId": "update_toolset_endpoint_api_v1_toolsets__toolset_id__put",
        "parameters": [
          {
            "name": "toolset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Toolset Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ToolSetUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ToolSet"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Toolset Endpoint",
        "description": "刪除工具集。\nDELETE /api/v1/toolsets/{toolset_id}",
        "operationId": "delete_toolset_endpoint_api_v1_toolsets__toolset_id__delete",
        "parameters": [
          {
            "name": "toolset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Toolset Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/agents/{agent_id}/toolsets": {
      "post": {
        "summary": "Assign Toolset To Agent Endpoint",
        "description": "分配工具集給 Agent。\nPOST /api/v1/agents/{agent_id}/toolsets",
        "operationId": "assign_toolset_to_agent_endpoint_api_v1_agents__agent_id__toolsets_post",
        "parameters": [
          {
            "name": "agent_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Agent Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentToolSetAssign"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentToolSetResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Get Agent Toolsets Endpoint",
        "description": "獲取 Agent 的所有工具集。\nGET /api/v1/agents/{agent_id}/toolsets",
        "operationId": "get_agent_toolsets_endpoint_api_v1_agents__agent_id__toolsets_get",
        "parameters": [
          {
            "name": "agent_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Agent Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/agents/{agent_id}/available-tools": {
      "get": {
        "summary": "Get Agent Available Tools Endpoint",
        "description": "獲取 Agent 可用的所有工具。\nGET /api/v1/agents/{agent_id}/available-tools",
        "operationId": "get_agent_available_tools_endpoint_api_v1_agents__agent_id__available_tools_get",
        "parameters": [
          {
            "name": "agent_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Agent Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/agents/{agent_id}/toolsets/{toolset_id}": {
      "delete": {
        "summary": "Remove Toolset From Agent Endpoint",
        "description": "移除 Agent 的工具集分配。\nDELETE /api/v1/agents/{agent_id}/toolsets/{toolset_id}",
        "operationId": "remove_toolset_from_agent_endpoint_api_v1_agents__agent_id__toolsets__toolset_id__delete",
        "parameters": [
          {
            "name": "agent_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Agent Id"
            }
          },
          {
            "name": "toolset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Toolset Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AgentToolSetAssign": {
        "properties": {
          "agent_id": {
            "type": "string",
            "title": "Agent Id",
            "description": "Agent ID"
          },
          "toolset_id": {
            "type": "string",
            "title": "Toolset Id",
            "description": "ToolSet ID"
          }
        },
        "type": "object",
        "required": [
          "agent_id",
          "toolset_id"
        ],
        "title": "AgentToolSetAssign",
        "description": "分配工具集給 Agent 的請求 Schema"
      },
      "AgentToolSetResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "agent_id": {
            "type": "string",
            "title": "Agent Id"
          },
          "toolset_id": {
            "type": "string",
            "title": "Toolset Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "agent_id",
          "toolset_id",
          "created_at"
        ],
        "title": "AgentToolSetResponse",
        "description": "Agent-ToolSet 關聯回應 Schema"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ToolSet": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Name",
            "description": "工具集名稱"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "工具集描述"
          },
          "tool_names": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tool Names",
            "description": "工具名稱列表"
          },
          "is_global": {
            "type": "boolean",
            "title": "Is Global",
            "description": "是否為全局工具集",
            "default": false
          },
          "id": {
            "type": "string",
            "title": "Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "name",
          "tool_names",
          "id",
          "created_at",
          "updated_at"
        ],
        "title": "ToolSet",
        "description": "工具集回應 Schema"
      },
      "ToolSetCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Name",
            "description": "工具集名稱"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "工具集描述"
          },
          "tool_names": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tool Names",
            "description": "工具名稱列表"
          },
          "is_global": {
            "type": "boolean",
            "title": "Is Global",
            "description": "是否為全局工具集",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "name",
          "tool_names"
        ],
        "title": "ToolSetCreate",
        "description": "創建工具集的請求 Schema"
      },
      "ToolSetUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "tool_names": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tool Names"
          },
          "is_global": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Global"
          }
        },
        "type": "object",
        "title": "ToolSetUpdate",
        "description": "更新工具集的請求 Schema（所有欄位都是可選的）"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}